perf: adapt commercial version push training data
