+++ projects/app/src/global/core/dataset/api.d.ts
@@ -27,13 +27,7 @@ export type CreateDatasetParams = {
 export type InsertOneDatasetDataProps = PushDatasetDataChunkProps & {
   collectionId: string;
 };
-export type PushDatasetDataProps = {
-  collectionId: string;
-  data: PushDatasetDataChunkProps[];
-  trainingMode: `${TrainingModeEnum}`;
-  prompt?: string;
-  billId?: string;
-};
+
 export type UpdateDatasetDataProps = {
   id: string;
   q?: string; // embedding content
